internal: move translation.d.ts to client
