+++ generators/client/translation.d.ts
