+++ generators/client/types.d.ts
@@ -4,8 +4,8 @@ import type { CypressApplication } from '../cypress/types.js';
 import type { JavaScriptApplication, JavaScriptSourceType } from '../javascript/types.js';
 import type { PostWritingEntitiesTaskParam } from '../../lib/types/application/tasks.js';
 import type { Language } from '../languages/support/languages.ts';
-import type { GetWebappTranslationCallback } from '../../lib/types/base/translation.js';
 import type { Entity } from '../base-application/index.js';
+import type { GetWebappTranslationCallback } from './translation.js';
 import type Command from './command.ts';
 
 type ApplicationClientProperties = ExportApplicationPropertiesFromCommand<typeof Command>;
